refactor: some cleanup and reordering 🎨
