+++ crates/handlers/src/lib.rs
@@ -1,9 +1,15 @@
 use sqlx::{any::AnyPoolOptions, AnyPool};
 use std::time::Duration;
 
-pub mod mysql;
-pub mod postgres;
-pub mod sqlite;
+mod handler;
+mod mysql;
+mod postgres;
+mod sqlite;
+
+pub use handler::{Handler, RowHandler, TableHandler};
+pub use mysql::MySQLHandler;
+pub use postgres::PostgresHandler;
+pub use sqlite::SQLiteHandler;
 
 pub async fn establish_connection(conn_string: &str) -> Result<AnyPool, String> {
     let pool = AnyPoolOptions::new()
