+++ crates/handlers/src/mysql/mod.rs
@@ -1,7 +1,7 @@
+use crate::handler::{Handler, RowHandler, TableHandler};
 use async_trait::async_trait;
 use serde_json::Value::{self as JsonValue};
 use sqlx::{any::AnyRow, AnyPool};
-use tx_lib::handler::{Handler, RowHandler, TableHandler};
 use tx_lib::types::{ColumnProps, FKRows, FkRelation};
 
 #[derive(Debug)]
