+++ crates/lib/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod decode;
 pub mod events;
 pub mod fs;
-pub mod handler;
 pub mod types;
